Attempting to fix docker build on push

+++ tests/build_tool.sh
@@ -51,4 +51,9 @@ fi
 # Cleanup
 rm $REPO/target/docker/burrow.dockerartefact
 rm $REPO/target/docker/burrow-client.dockerartefact
-docker rmi -f $IMAGE:build
+
+# CircleCI seems to have an issues removing this build, in any case it is not necessary on CI
+if [ ! "$CI" ]
+then
+  docker rmi -f $IMAGE:build
+fi
